+++ NEWS.md
@@ -1,5 +1,7 @@
 # valr (development version)
 
+* Added `bed_genomecov()` to compute interval coverage across a genome.
+
 # valr 0.7.0
 
 * `read_bed` and related functions now automatically calculate the fields. Use of `n_fields` was deprecated.